pyproject.toml
[project]
name = "file-downloader"
version = "0.1.1"
description = "File downloader"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aria2p>=0.12.1",
    "paho-mqtt>=2.1.0",
    "toml>=0.10.2",
]

[project.scripts]
fetcher = "fetcher:main"
puller = "puller:main"

[tool.uv]
package = true
